monogame-patcher: Work around terminal width issue
If the command is not executed inside a terminal the value of
Console.WindowWidth is 0 and thus we will get an exception from
CommandLineParser:

Unhandled Exception:
System.ArgumentOutOfRangeException: Length cannot be less than zero.
Parameter name: length
  at System.String.Substring ...
  at CommandLine.Text.HelpText.AddOption ...
  ...

So let's initialize the parser with settings and if WindowWidth is 0 we
just add 80 as the width.

Signed-off-by: aszlig <aszlig@nix.build>