monogame-patcher: Fix assembly search directory
By default the assemblies are searched in the current working directory,
but when patching an assembly in some other directory than the current
the patcher will fail to resolve assemblies residing beneath the target
file.

So we just add the directory of the target file to the search path.

I also moved all of the assemblies in the test to be in a subdir so we
can verify that this indeed works and we won't regress in the future.

Signed-off-by: aszlig <aszlig@nix.build>