monogame-patcher: Allow to define specific methods
So far we only matched a substring of a type, which isn't very exact of
course. We also weren't able to specify the method to patch.

Now the methods that are being patched can be specified by using
Type::Method. If it's just Type without a method all the methods of the
type and its subtypes are patched.

I also added a small check to make sure that all of these methods were
found, and if not an error is thrown.

Signed-off-by: aszlig <aszlig@nix.build>