games/gog: Switch everything to gogUnpackHook
For Albion, Settlers 2 and World of Xeen we need to explicitly add
gogUnpackHook, because for those the derivation containing the game data
is not created via our buildGame wrapper.

Everything else is just done by removing unpackCmd and unzip from the
package arguments. One exception is Thimbleweed Park, which relied on
custom path for unzip in order to get the desktop icon, but with our now
hook, we get that icon now in xdg-icon.png with a flat directory
structure.

I've tested this by building all the games from GOG that we have
packaged here. However, I didn't test whether every single one actually
runs.

Signed-off-by: aszlig <aszlig@nix.build>