games: Move FMOD fixing from Bastion to setup hook
Thanks to @layus for pointing out that there is at least one other game
(Epistory) which has the same FMOD issue as Bastion has.

So I decided to move this into a setup hook that automatically discovers
whether it's the affected FMOD version and NOPs out the calls to
system().

In summary: If another game is affected, all that's needed now is to add
fixFmodHook to nativeBuildInputs.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @layus