pkgs/vim: Update Nix plugin
This is mainly to catch up with the syntax highlighting fixes that were
done in the respective plugins.

The setlocals in ftplugin/nix.vim from LnL7's plugin are now gone and
it's one setlocal with line continuation, so I had to turn the simple
grep into a small sed script.

Another thing I added while at it is to add the single quote to
iskeyword, because it can be part of a Nix identifier. I did that by
replacing the dash, because the latter now is part of the setlocal of
the aforementioned ftplugin.

Signed-off-by: aszlig <aszlig@nix.build>