pkgs/sandbox: Add dependency on Boost
We're not directly using Boost but some headers are included in the
headers we use, which causes the build to fail on newer Nix versions.

In theory, this should be propagated by the Nix derivation. See also
this comment:

https://github.com/NixOS/nixpkgs/issues/39001#issuecomment-381949306

However, when looking at the Nix derivation, it already contains code
that takes care of not propagating the boost dependency.

Nevertheless, adding the dependency on our end even though it could
become redundant some day certainly won't hurt.

Signed-off-by: aszlig <aszlig@nix.build>