tests/bfq: Don't test on bleeding edge kernel
If someone is going to use our patch for the latest release canidate
kernel and it doesn't work I'd say it's somewhat expected.

Right now the test fails with the current 4.19-rc6 and I don't have time
to investigate this. It might be the case that this also might be an
upstream issue, but let's actually lazy evaluate this and re-check once
4.19 is released.

Signed-off-by: aszlig <aszlig@nix.build>