profiles/services: Remove erroneous rngd
Being tired and trying to do Nix doesn't work, so removing
`services.rngd` because it doesn't exist (it's security.rngd). Thinking
about it, there's probably no reason to explicity enable it if it's
already enabled by default because at the end of the day we can still
look it up if needed.