tests/gnupg: Fix OCR match on passphrase dialog
Sinc the Tesseract update to version 4.0.0[1], the OCR result is much
better than what we had before.

The string we want to match is the following:

  Please enter the passphrase to unlock the OpenPGP secret key

In the previous Tesseract alpha version (4.00.00alpha-git-20170410), the
word "passphrase" was detected with an upper-case "P", which now is no
longer the case.

Nevertheless, I've changed it to matching both "Passphrase" and
"passphrase", just to be sure.

[1]: https://github.com/NixOS/nixpkgs/commit/d4b9752212f77b122b9166dc52f

Signed-off-by: aszlig <aszlig@nix.build>