games/crosscode: Update to version 1.0.2
This update contains mostly Steam-related changes, which are not
relevant to the GOG version.

However here are the changes that affect us:

  * The plaza of the Last Minute Heroes HQ can be entered and it
    includes the first few statues! The interior is still closed off,
    but it will open up with the next update.
  * There are a few more NPCs in CrossWorlds including a new cameo guest
    role somewhere in Bergen village.
  * The new stat "Time played with Assist Mode" was added to the
    Statistics menu.

The full upstream blog post can be found at:

http://www.radicalfishgames.com/?p=6520

With this update a bunch of different Greenworks[1] versions are now in
a new "modules" subdirectory, which I removed because they contain
native code for interfacing with Steamworks. This doesn't apply to us,
so let's actually make sure that this code never runs by removing the
modules directory entirely.

[1]: https://github.com/greenheartgames/greenworks

Signed-off-by: aszlig <aszlig@nix.build>