machines/dnyarri: Enable bcache kernel module
Since I got a new SSD for the machine (thanks @cvdnext), I also had the
opportunity to re-create my LUKS containers to LUKS2 with Argon2 key
derivation alongside creating bcache backing devices.

The change in order to support bcache is just a matter of adding
"bcache" to availableKernelModules and we're done.

However, as the storage configuration is not a very common one, I
decided to add a test specific to that to make sure future NixOS updates
won't prevent the machine from booting.

Signed-off-by: aszlig <aszlig@nix.build>