fetchHumbleBundle: Fix getting recaptcha response
The object in question to get the response from the Recaptcha widget is
now called "grecaptcha" instead of just "captcha" and the method to call
now is also called "getResponse" instead of "get_response".

Another change is that the grecaptcha object isn't available immediately
once the page is loaded, so I added a try...catch loop for that. I know
it's evil to do that, but I tried to check just for whether grecaptcha
is available and then call the getResponse method but found out that
during object construction there is a small time window where
getResponse is available but isn't a callable method.

There is still another issue with the HB fetcher, but the first step is
hereby fixed.

Signed-off-by: aszlig <aszlig@nix.build>