games: Use "mono" instead of "mono50"
The use of mono50 was just an old relic where I was debugging some issue
with a game, but I used it as a template for packaging other games, so
it spread throughout the code base.

I've tested all these games just to make sure they still work with the
latest version.

Signed-off-by: aszlig <aszlig@nix.build>