monogame-patcher: Fix SHA256 hashes for deps
I checked the archives against the previous versions and the contents
match, however the archives themselves had differences with the file
order so the hashes are different now.

In the long term, we probably should use fetchzip instead so we don't
get a hash mismatch if none of the actual contents change.

Signed-off-by: aszlig <aszlig@nix.build>