modules/vlock: Allow to change the user name
When writing that module back then in 88a10f1a6e2ea172ff985f64bfcf82e2,
I had to work around a systemd limitation (which I need to check whether
this is still the case today) and hardcoded the user name for the time
being.

While it's still quite ugly, there was interest from @Profpatsch, so I
decided to add an internal option to change that username.

Signed-off-by: aszlig <aszlig@nix.build>