fetch-gog: Provide a better login failure message
So far if the login credentials were wrong, you get something like this:

  Traceback (most recent call last):
    File "/nix/store/...-fetch-gog.py", line 113, in <module>
      GogFetcher(sys.argv[1], sys.argv[2], sys.argv[3]).fetch()
    File "/nix/store/...-fetch-gog.py", line 15, in __init__
      self.login()
    File "/nix/store/...-fetch-gog.py", line 28, in login
      auth_code = parse_qs(urlsplit(browser.get_url()).query)['code']
  KeyError: 'code'

This is isn't very helpful and might hint that there is something wrong
with the fetcher. Of course, if GOG would change their login interface
and no longer expose the "code" query string item, it might still say
login failure, but right now we don't have a better way to detect it.

At least if we get a login failure with a login that has been working so
far, we know something has changed upstream :-)

Signed-off-by: aszlig <aszlig@nix.build>
Reported-by: @Profpatsch