games/gog: Add Heroes of Might & Magic 3: Complete
I had this laying around for quite a while but haven't committed it yet,
so I decided to update VCMI to the latest master version.

Instead of using the "mp3" directory in the game assets, I decided to go
for "music" instead, because all of the files within that directory are
actually Ogg Vorbis and it's the game's music.

Another thing I've changed is that the launcher now uses execl() instead
of QProcess::startDetached, because the latter doesn't work very well
with our sandbox and will subsequently cause the client to be killed.

Signed-off-by: aszlig <aszlig@nix.build>