games: Use preferLocalBuild for fetchers
If we're on a machine with remote builds enabled, we'd like to prefer
not having our login credentials to be pushed on a remote server.

We of course also don't want the actual games to end up on a different
machine, so let's use preferLocalBuild to avoid this.

Unfortunately avoidance is not at any cost, because if we for example
use a different system attribute, it might be built on a remote machine
nevertheless. Ideally there would be something like "forceLocalBuild",
which avoids remote builds at any cost.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch