games/gog: Add The Longest Journey
This is basically a wrapper around ResidualVM, which I also patched so
that we can inject a basic config file from the Nix store that contains
the basic game information.

The ResidualVM build is pretty minimal in that we only use those
dependencies that we actually need for the game.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch