machines/brawndo: Add cnijfilter2 version 5.70
I don't recall the exact model which was used for that machine, but it
was definitely one that required version 5.70.

Now I made a pull request for nixpkgs a while ago[1], which hasn't been
merged and as I didn't have the chance to test it by myself I didn't
merge that pull request yet.

Since the user asked about that printer driver again, I decided to give
version 5.70 a shot in vuizvui only and merge it upstream, once the user
reports that the driver is working.

[1]: https://github.com/NixOS/nixpkgs/pull/58399

Signed-off-by: aszlig <aszlig@nix.build>