pkgs/aszlig: Add custom Psi XMPP client
This is from the current 2.0 development version and has most of my own
configuration preferences built-in.

Right now, the theming is pretty much a work in progress and the chat
input currently shows black text on dark grey background, which is quite
a nuisance to use.

Another thing that's currently not working (or just for a very short
amount of time) are window manager urgency hints.

Nevertheless however, I'm already using it as my main XMPP client
despite these issues.

Signed-off-by: aszlig <aszlig@nix.build>