{machines,modules}/devhell: Split offlineimap
I can't keep `offlineimap` as a service that applies to all, simply
because `gunnr` will never be able access IMAP from its network.
There's therefore little point in having it try to connect. So, instead
I've moved the service to all machines that are expected to be able to
access emails.