machines/dnyarri: Fix allowUnfreePredicate
Since NixOS/nixpkgs@c814d72b517bb201c8bbbfc64e386c7023352886, a lot of
packages now no longer have a name attribute but instead use pname, so
when checking the package name within allowUnfreePredicate we need to
make sure that we fall back to a default if the name attribute is not
present.

This fixes evaluation of the machine, however I didn't check if it
actually builds.

Signed-off-by: aszlig <aszlig@nix.build>