modules/zsh: Fix options and clean up a bit
Since a while ago[1], the setting of ZSH options is now done after
interactiveShellInit, so using unsetopt SHARE_HISTORY doesn't work
anymore because it is set *afterwards*.

Instead of setting these options, we now use the setOptions option
instead and override it with exactly the options I want to be set.

Additionally, compinit is also no longer necessary, because it is done
by default and invoking it on our own is just redundant.

[1]: https://github.com/NixOS/nixpkgs/pull/58012

Signed-off-by: aszlig <aszlig@nix.build>