devhell/profiles: Fix VirtualBox kernel 5.3 build
This uses a patch from Arch Linux to fix the build of the VirtualBox
kernel modules for Linux 5.3. I didn't do any testing of this other than
just to build the modules with the fix.

The override here is to be reverted as soon as we got a fix for this in
upstream nixpkgs. I didn't do this yet by myself, because I *really*
should get back to work right now...

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @devhell