fetch-humble-bundle: Fix login with guard code
HumbleBundle has introduced a new guard code, which is sent to the email
address of the account owner after the captcha has been solved.

Unfortunately, the humblebundle-python[1] library is no longer
maintained and there are no forks with fixups. In addition, the code of
the library is sub-par and also doesn't support Python 3.

So I added a patch containing a _very_ dirty fix to include the guard
code during the login procedure for now until we have decided to either
go with the existing library and fork it or use another library or
implement it ourselves, since we only use a tiny fraction of the HB API.

I also fixed up the Qt plugin path for the captcha prompt if a newer
nixpkgs revision is used, where we have wrapQtAppsHook.

One downside of the guard code is that we get prompted for it _every_
time we download something, which might be annoying, but at least it's
better than not working at all.

Signed-off-by: aszlig <aszlig@nix.build>