fetch-gog: Fix login with captcha
The way GOG embeds recaptcha has changed a while ago and it's now
embedded via additional JavaScript which no longer directly contains
"google.com/recaptcha", because it is now contained within an iframe.

Fortunately, the fix is relatively easy, because everything else is
still implemented in the same way, we just need to match portions of
that JavaScript code.

I also fixed up the Qt plugin path for the captcha prompt if a newer
nixpkgs revision is used, where we have wrapQtAppsHook.

Signed-off-by: aszlig <aszlig@nix.build>