profiles/base: Fix networking.useDHCP clash
This addresses a new assertion to the way NixOS handles networks with
networkd [1].

As a result this disables `networking.useDHCP` globally on all my
machines. Explicit interface configuration, as requested by [1] will be
handled in separate commits due to separate machines, not all of which I
have access to right now.

[1] https://github.com/NixOS/nixpkgs/pull/69302