machines/aszlig: Explicitly set useDHCP
The usage of DHCP is no longer global since a while[1] and we now have
to explicitly enable it for the interfaces in question.

This actually is a good thing and makes it far less problematic if we
use tunnel interfaces and other more complicated networking
configuration.

I added the definitions for all machines where I actually know which
interfaces are in use and disabled useNetworkd for shakti, because I
don't know the interface names for that machine and the machine
currently isn't in use anyway, so we can add it later if needed.

[1]: https://github.com/NixOS/nixpkgs/pull/69302

Signed-off-by: aszlig <aszlig@nix.build>