profiles/workstation: Add wcc to packages
This is the "Witchcraft Compiler Collection", which is VERY useful for
reverse-engineering, especially when looking back at the work I've done
with game packaging here in Vuizvui, I would have needed something like
this a ton of times.

I've mainly used radare2 for dissecting the binaries, but especially the
"Witchcraft Shell" is quite useful to have.

Thanks to @Profpatsch for pointing this out to me.

Signed-off-by: aszlig <aszlig@nix.build>