invisigun-heroes: Update to version 1.7.16
Lots of changes have been introduced in version 1.7.6, so here is a
small summary:

  * New "The Hero's Journey" mode, which allows to discover the
    abilities of each hero along with challenges and puzzles.
  * Three new heroes: Djaan-Khe, Zephyr, Violet
  * Challenge mode has been removed (replaced by "The Hero's Journey").
  * Guest Edition has been removed.

The full changelog can be found here:

http://www.sombr.com/pub/SIG/changelog/

While Invisigun Heroes in general went pretty smooth when it comes to
packaging on our side, this time unfortunately we needed to patch
something.

There is now a new "config.json", which is written inside the game's
data directory. Of course, since the data directory resides in the Nix
store, I've relocated the file to be written into the the config
directory ($XDG_CONFIG_HOME/unity3d/Sombr Studio/Invisigun Reloaded) of
the game.

Note that the config directory has also changed from "Invisigun Heroes"
to "Invisigun Reloaded", so if you don't want to play the old version,
you can delete the "Invisigun Heroes" directory.

Signed-off-by: aszlig <aszlig@nix.build>