pkgs/psi: Update to latest upstream master
Updating to latest master fixes one of the most annoying theming issue
where the chat message input box text will turn black on dark background
for my custom Qt theme.

There are also quite a lot of fixes in master (which is still
unreleased), which we want to have as well.

Signed-off-by: aszlig <aszlig@nix.build>