pkgs/psi: Temporarily disable Jingle sessions
When getting an incoming file transfer, Psi has crashed a while ago and
I unfortunately don't remember the exact details. However, since those
stanzas are going to be re-delivered every time we start Psi again, I'd
like to debug this in an isolated environment rather than being more or
less forced to stay unavailable (especially annoying if someone is
trying to send a file and then you stay offline for hours).

So until I get to debugging this and ideally also test this in an
automated fashion, I'll disable it for now since I'm not using Jingle
sessions at the moment.

Signed-off-by: aszlig <aszlig@nix.build>