monogame-patcher: Also patch late-bound methods
This is something I encountered in Invisigun Heroes, where there is a
virtual FileManagerAdapter class, which then gets bound to either the
Desktop variant or any of the other ones.

Since we always use the full path of the corresponding type/object, just
adding the Callvirt to the opcodes we patch shouldn't interfere with any
other games we patch that way and using an extra subcommand just for the
Callvirt case would only add additional boilerplate for no additional
benefits.

Signed-off-by: aszlig <aszlig@nix.build>