machines/hildr: Add exceptions for certain ifaces
This has come about because qemu-kvm VMs lost the ability to gain a DHCP
issued IP address after a reboot. As it turns out networkd seems to have
played a part in messing with the virbr0 interface. Thanks to @aszlig we
figured out what was going on, so now we're going to make networkd
ignore all interfaces that are not physical.