vim: Remove file artifacts for ATS syntax plugin
We essentially only want to have the ftdetect and syntax files,
everything else is just cruft that is not in any way related to Vim.

Ideally we want to do a whitelist approach instead of the "remove
everything unneeded" we're doing right now, but since I don't want to
refactor the whole Vim expression I'll leave it the dirty[TM] way.

Signed-off-by: aszlig <aszlig@nix.build>