vim: Add vim-css-color plugin
This is quite useful in CSS files and others to directly highlight the
colours in the actual colour value rather than with the generic colour
of the syntax file.

To make sure we don't break the after/syntax files for Haskell and HTML,
I also changed the way we install those files in the output directory so
that if a file already exists, it is appended to rather than
overwritten.

Signed-off-by: aszlig <aszlig@nix.build>