pkgs/aszlig/xournal: Switch to xournal++
The patch I added to xournal was for keeping the aspect ratio when
annotating PDFs with images. However, looking at xournal++ the aspect
ratio is kept by default when resizing via corners so the patch is not
needed.

Since I don't really care a lot whether it's xournal or xournal++ and as
long as it does the very little things I intend to use it for, I don't
mind if it has too many features for my taste.

Signed-off-by: aszlig <aszlig@nix.build>