profpatsch/pkgs: Switch to using wrapMpv
Recently, the way to add scripts to the mpv wrapper has changed[1] and
instead of using .override, there is now a dedicated wrapMpv function
that can be used to compose the wrapped mpv derivation, similar on how
it's done for wrapFirefox and wrapNeovim.

The change also introduced the following evaluation error when using the
old mpv-with-scripts package:

  Use wrapMpv for editing the environment of mpv

Since this evaluation error essentially blocks Vuizvui channel
generation, I decided to fix this, even though I usually try to avoid
touching other people's machine configurations.

[1]: https://github.com/NixOS/nixpkgs/commit/f93918bdc387d353285f458c06c6a111ae90b7b2

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch