core/tests: Fix attribute paths for wireguard
A few days ago, the wireguard tests were consolidated[1] into a single
attribute with various subattributes, similar to how the virtualbox
tests and many other tests are structured.

I fixed the attributes and also added the wg-quick attribute that was
also introduced in the meantime[2].

This should fix the evaluation errors occuring whenever a channel
contains a machine configuration with wireguard enabled.

[1]: https://github.com/NixOS/nixpkgs/commit/41bd6d2614749d12ce5ded3e991555b56ea6b2dc
[2]: https://github.com/NixOS/nixpkgs/commit/abf60791e2bd274d39e0f18def46795798f9aefd

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch