games/gog: Add Freedom Planet version 1.21.5
A game that I had laying around since quite a while but I was too lazy
to patch properly, since the game expects its data files as well as its
save files in the current working directory.

While I did patch the game via an LD_PRELOAD wrapper of fopen it also
feels kinda rendundant with code we have in preloaders of other games.

So in the long term we might want to implement something a bit more
generic, but for now the game works and config and saves are properly
placed in XDG_CONFIG_HOME and XDG_DATA_HOME.

Signed-off-by: aszlig <aszlig@nix.build>