setup-hooks/fix-fmod: Patch all calls to system()
In newer versions of libfmod, there are two calls to system(), which
cause the setup hook to fail, since it only expects one such call.

Fortunately, we don't want libfmod to execute *any* external commands,
so fixing this is rather easy by just making *all* calls to system() a
no-op.

Signed-off-by: aszlig <aszlig@nix.build>