legosi: Use modulesPath instead of <nixpkgs/nixos>
While it's fine to use <nixpkgs> on most systems, we deliberately want
to avoid the use of <nixpkgs> to make sure that whenever we for example
run <nixpkgs/nixos/lib/eval-config.nix> with a custom "pkgs" argument we
are guaranteed that we get the version we specify.

So this is one of the reason I used <nixpkgsSrc> on Hydra instead of
<nixpkgs>, so that whenever we have such occasions where we can't
guarantee such things, the evaluation will fail.

And right now, it does:

  in job 'machines.profpatsch.legosi':

  file 'nixpkgs/nixos/modules/profiles/qemu-guest.nix' was not found in
  the Nix search path (add it using $NIX_PATH or -I), at
  .../machines/profpatsch/legosi.nix:12:5

Fortunately, there is modulesPath, which refers to
<nixpkgs/nixos/modules> of the nixpkgs version passed via "pkgs".

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @Profpatsch