fetchGog: Default downloadName to "en3installer0"
Most of the games are using "en3installer0" as the downloadName, so it
makes sense to use this as the default, since first of all do not
support languages other than English except for a few games (eg. Albion)
and second, there is no game with several installer packages to chose
from (at least not on top of my head).

Even if there would be, it would still be the exception rather than the
rule, so let's optimise for the common scenario.

Signed-off-by: aszlig <aszlig@nix.build>