sandbox: Use closureInfo for getting runtime deps
A while ago[1], @Profpatsch added this comment above the definition of
the $runtimeDeps variable:

  # Reads the dependency closures and does ? something? TODO: explain

I just recently (yesterday as of the date of this commit) found out
about that comment by accident.

While this should probably be better of as an issue instead, the comment
does have a point, since not everybody enjoys reading/writing sed
expressions.

In a nutshell, what the implementation actually does is parsing the
output of the files generated by exportReferencesGraph.

At the time of writing the implementation, we didn't have a JSON-based
interface in Nix for doing the same, nor did we have something like
pkgs.closureInfo. There was only a small Perl script[2], which did
something like this, but given that it can be easily done via sed, I
opted to instead use the latter.

Nevertheless however, using closureInfo is not only more concise in its
implementation, it also makes our implementation much more concise as
well and also obvious on what we're doing here.

[1]: 09dc1d8ad625b9a1d5b89593b184d316837ba1cc
[2]: https://github.com/NixOS/nixpkgs/blob/8747190024205a5a3534b4e9a18dbaf3f3ee7b39/pkgs/build-support/kernel/paths-from-graph.pl

Signed-off-by: aszlig <aszlig@nix.build>